- Move algorithm implementation to project.
